Rename new-homepage to home
