+++ components/home/CollaborateWithMoneySection.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 
+import NextIllustration from '../collectives/HomeNextIllustration';
 import { Box, Flex } from '../Grid';
-import NextIllustration from '../home/HomeNextIllustration';
 import { SectionDescription, SectionTitle } from '../marketing/Text';
 import { H5, P } from '../Text';
