+++ components/home/DedicatedTeamSection.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 
+import NextIllustration from '../collectives/HomeNextIllustration';
 import Container from '../Container';
 import { Box, Flex } from '../Grid';
-import NextIllustration from '../home/HomeNextIllustration';
 import Link from '../Link';
 import { SectionDescription, SectionTitle } from '../marketing/Text';
 import StyledButton from '../StyledButton';
