+++ pages/home.js
@@ -1,16 +1,13 @@
 import React from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 
-import Banner from '../components/home/Banner';
-import CreateCollective from '../components/home/sections/CreateCollective';
-import FeaturesSection from '../components/home/sections/Features';
-import FiscalHostSection from '../components/home/sections/FiscalHost';
-import JoinUsSection from '../components/home/sections/JoinUs';
-import LearnMoreSection from '../components/home/sections/LearnMore';
-import MakeCommunitySection from '../components/home/sections/MakeCommunity';
-import OCUsersSection from '../components/home/sections/OCUsers';
-import WeAreOpenSection from '../components/home/sections/WeAreOpen';
-import WhatCanYouDoSection from '../components/home/sections/WhatCanYouDo';
+import JoinUsSection from '../components/collectives/sections/JoinUs';
+import CollaborateWithMoney from '../components/home/CollaborateWithMoneySection';
+import DedicatedTeam from '../components/home/DedicatedTeamSection';
+import GetToKnowUs from '../components/home/GetToKnowUsSection';
+import OpenCollectiveIs from '../components/home/OpenCollectiveIsSection';
+import RaiseMoney from '../components/home/RaiseMoneySection';
+import TheFutureIsCollective from '../components/home/TheFutureIsCollectiveSection';
 import Page from '../components/Page';
 
 const menuItems = { pricing: true, howItWorks: true };
@@ -22,25 +19,22 @@ const messages = defineMessages({
   },
 });
 
-const HomePage = () => {
+const NewHomePage = () => {
   const { formatMessage } = useIntl();
   return (
     <Page menuItems={menuItems} description={formatMessage(messages.defaultTitle)}>
-      <Banner />
-      <MakeCommunitySection />
-      <WhatCanYouDoSection />
-      <FeaturesSection />
-      <FiscalHostSection />
-      <CreateCollective />
-      <OCUsersSection />
-      <WeAreOpenSection />
-      <LearnMoreSection />
+      <TheFutureIsCollective />
+      <RaiseMoney />
+      <OpenCollectiveIs />
+      <CollaborateWithMoney />
+      <DedicatedTeam />
+      <GetToKnowUs />
       <JoinUsSection />
     </Page>
   );
 };
 
-HomePage.getInitialProps = ({ req, res }) => {
+NewHomePage.getInitialProps = ({ req, res }) => {
   if (res && req && (req.language || req.locale === 'en')) {
     res.set('Cache-Control', 'public, s-maxage=3600');
   }
@@ -55,4 +49,4 @@ HomePage.getInitialProps = ({ req, res }) => {
   return { skipDataFromTree };
 };
 
-export default HomePage;
+export default NewHomePage;
